init: update recovery when enabled in settings
Update the recovery image only if the option is enabled
under Developer options

This reverts commit 231e0a9e6a1da6fa4a188840f68af649669e417f.

Change-Id: I928f7ee8bb3eaf5581bb8225661d253ecca0c4ef

Change CM recovery install script filename [2/2]

This is part 2/2 to maintain backwards compatibility with CWM's
verify_root_and_recovery() function. CWM checks if install-recovery.sh
exists and has an executable flag set, then offers to disable the script
for the user. CM now controls this with the persist.sys.recovery_update
property which is configurable via settings, so we don't need to
double-check this.

This changes the name of the recovery install script to
install-cm-recovery.sh.

init: Move install-recovery.sh back to the standard location

L moved the location of install-recovery.sh from /system/etc/ to
/system/bin. Since CWM recovery isn't looking for this location
anyway, let's return the file to this standard location. This allows
all other code in L to function properly.

Maintain the change to the init to allow flash_recovery to be disabled
in settings.

Change-Id: I8a85db8addeb75a2fd60d809c5ed4edc619ef7ed

Former-commit-id: 770c2eff9be48e0d2a0695b7b133c73cadf506f3
Former-commit-id: 534c62cc1291041ca66a248f763235531f7bf36e
Former-commit-id: 2d465f0cf69bc82f0c2c993ee797de6037fa1d48
Former-commit-id: a3c8e2b494a1319aa43851bc6a545933089fa1be
Former-commit-id: 329d1a39580072177c346a8cee90f6b9334819cb
Former-commit-id: 97142c25ff28b6c0a4a916404225c064a977bf6b

Change-Id: I275dd358b46c626dfcb8fe02c583a308d5a89c56